requirements.txt
requests
discord-py-interactions